+++ src/raster/bins/tests.rs
@@ -18,6 +18,8 @@ fn triangle(x0: i32, y0: i32, x1: i32, y1: i32) -> Prepared {
             y: y * SUBPIXEL_SCALE,
         },
         z: 1 << 31,
+        s: 0,
+        t: 0,
     };
     prepare([p(x0, y0), p(x0, y1), p(x1, y0)], 0).expect("triangle visible")
 }